get, post, delete tag controllers

+++ src/main/java/com/skillbox/javapro21/service/impl/TagServiceImpl.java
@@ -1,9 +1,9 @@
 package com.skillbox.javapro21.service.impl;
 
+import com.skillbox.javapro21.api.response.DataResponse;
 import com.skillbox.javapro21.api.response.ListDataResponse;
 import com.skillbox.javapro21.api.response.tag.TagData;
 import com.skillbox.javapro21.domain.Tag;
-import com.skillbox.javapro21.exception.BadArgumentException;
 import com.skillbox.javapro21.repository.TagRepository;
 import com.skillbox.javapro21.service.TagService;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +14,9 @@ import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Component
 @RequiredArgsConstructor
@@ -31,20 +33,59 @@ public class TagServiceImpl implements TagService {
         return getTagResponse(pageableTagList, offset, itemPerPage);
     }
 
-    public DataResponse<TagData> deleteTagById(long id) throws BadArgumentException {
-        Tag tag = tagRepository.findById(id).orElseThrow(() -> new BadArgumentException("Тег с id " + id + " не найден"));
+    public DataResponse<TagData> addTag(String name){
+        DataResponse<TagData> dataResponse = new DataResponse<>();
+        TagData tagData = new TagData();
+        Tag tag = tagRepository.findByName(name);
+        if (tag == null){
+            tag = new Tag();
+            tag.setTag(name);
+            dataResponse.setError("ok");
+            dataResponse.setTimestamp(utilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
+            dataResponse.setData(tagData);
+            tagRepository.save(tag);
+            tagData.setTag(name);
+            return dataResponse;
+        }
+        dataResponse.setError("ok");
+        dataResponse.setTimestamp(utilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
+        tagData.setTag(tag.getTag());
+        tagData.setId(tag.getId());
+        dataResponse.setData(tagData);
+        tagRepository.save(tag);
+        return dataResponse;
+
+    }
+
+    public DataResponse<TagData> deleteTagById(long id) {
+        DataResponse<TagData> dataResponse = new DataResponse<>();
+        Tag tag = getTagById(id);
+
+        if (tag == null){
+            dataResponse.setError("invalid_request");
+            dataResponse.setTimestamp(utilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
+            return dataResponse;
+        }
         TagData tagData = new TagData()
                 .setId(tag.getId())
                 .setTag(tag.getTag());
         tagRepository.delete(tag);
 
-        return utilsService.getDataResponse(tagData);
+        dataResponse.setError("ok");
+        dataResponse.setTimestamp(utilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
+        dataResponse.setData(tagData);
+
+        return dataResponse;
+    }
+
+    private Tag getTagById(long id){
+        return tagRepository.findById(id).orElse(null);
     }
 
     private ListDataResponse<TagData> getTagResponse(Page<Tag> tags, int offset, int itemPerPage){
         ListDataResponse<TagData> listDataResponse = new ListDataResponse<>();
         listDataResponse.setError("ok");
-        listDataResponse.setTimestamp(UtilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
+        listDataResponse.setTimestamp(utilsService.getTimestampFromLocalDateTime(LocalDateTime.now()));
         listDataResponse.setTotal((int) tags.getTotalElements());
         listDataResponse.setOffset(offset);
         listDataResponse.setPerPage(itemPerPage);
@@ -66,6 +107,24 @@ public class TagServiceImpl implements TagService {
         return new TagData()
                 .setId(tag.getId())
                 .setTag(tag.getTag());
+    }
+
+    public Set<Tag> addTagsToPost(String[] tagsString) {
+        Set<Tag> tags = new HashSet<>();
+        for (String tagFromNewPost : tagsString
+        ) {
+            Tag tag = tagRepository.findByName(tagFromNewPost);
+            if (tag != null) {
+                tags.add(tag);
+            } else {
+                Tag newTag = new Tag();
+                newTag.setTag(tagFromNewPost);
+                tagRepository.save(newTag);
+                tags.add(newTag);
+            }
+        }
+
+        return tags;
     }
 
 }
